+++ frontend/src/components/ProductsDisplay.tsx
@@ -1,5 +1,7 @@
 'use client'
 
+import Link from 'next/link';
+
 import formatMoney from '@/utils/formatMoney';
 import { useMutation } from '@apollo/client';
 import { useUser } from './User';
@@ -57,16 +59,28 @@ export default function ProductsDisplay({
                 <span key={tag.name}>{tag.name}</span>
               ))}
             </p>
-            <div className={styles.buttonWrapper}>
-              <fieldset disabled={loading}>
-                <button type="button">Edit</button>
-                <button type="button" onClick={() => addItemToCart(product)}>Add</button>
-                <button type="button">Delete</button>
-              </fieldset>
-            </div>
+              <div className={`${styles.buttonWrapper} ${!user ? styles.loggedOut : ''}`}>
+                  <fieldset disabled={loading}>
+                    {user && (
+                      <>
+                        <button type="button">Edit</button>
+                        <button type="button" onClick={() => addItemToCart(product)}>Add</button>
+                        <button type="button">Delete</button>
+                      </>
+                    )}
+                    {!user && (
+                      <Link
+                        href="/signin"
+                        className="button"
+                      >
+                        Sign in to purchase
+                      </Link>
+                    )}
+                  </fieldset>
+              </div>
           </div>
-        )
-      })}
+        )})
+      }
     </div>
   );
 }